Changing sprite source in a running game.
Currently if one want to change a sprite `source` in a running game it will have to be done manually i.e changing source requires to change sourceX, sourceY, sourceWidth, sourceHeight, width, height and also we will need to remember the type of source ... It will be good to have a method like `setTexture` to do all this. A small change to the existing code is required.